refactor: move source code and tests into ./src
