+++ jest.config.js
@@ -1,3 +1,8 @@
 module.exports = {
   setupFilesAfterEnv: ['./jest.setup.js'],
+  testPathIgnorePatterns: [
+    '<rootDir>/build/',
+    '<rootDir>/node_modules/',
+    '<rootDir>/src/__tests__/mocks/',
+  ],
 };
