+++ src/__tests__/deleteImage.js
@@ -1,4 +1,4 @@
-const imgur = require('../lib/imgur.js');
+const imgur = require('../imgur.js');
 
 describe('deleteImage()', () => {
   describe('delete image response', () => {
