+++ src/__tests__/favoriteImage.js
@@ -1,4 +1,4 @@
-const imgur = require('../lib/imgur.js');
+const imgur = require('../imgur.js');
 
 describe('favoriteImage()', () => {
   describe('favorite image response', () => {
