+++ src/__tests__/mocks/handlers/image.js
