+++ src/__tests__/updateInfo.js
@@ -1,4 +1,4 @@
-const imgur = require('../lib/imgur.js');
+const imgur = require('../imgur.js');
 
 describe('updateInfo', () => {
   describe('update image metadata', () => {
